<sede> develop: arreglo en reportes de compras, Corregir sumas de reportes

+++ netbeans/sede/src/main/webapp/js/jqueryUtil.js
@@ -68,7 +68,7 @@ function sumaColumnaWithoutFormat(evento, clase, idViewTotal) {
 
     for (i = 0; i < numeroCeldas; i++) {
         let valor = $(celdas[i]).val();
-        sumaTotal += parseInt(valor);
+        sumaTotal += parseFloat(valor);
     }
     $("#" + idViewTotal).val(sumaTotal);
     return sumaTotal;
